Relax tightness of coupling in monitor code (#174)
* First move

* Clean up V1

* Reorganized code a bit

* Fix flake8 issue, tried to test lack of uamqp
