+++ azext_iot/monitor/handlers/__init__.py
@@ -1,0 +1,9 @@
+# coding=utf-8
+# --------------------------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for license information.
+# --------------------------------------------------------------------------------------------
+
+from azext_iot.monitor.handlers.handler import CommonHandler
+
+__all__ = ["CommonHandler"]
