+++ azext_iot/monitor/handlers/base_handler.py
@@ -1,0 +1,16 @@
+# coding=utf-8
+# --------------------------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for license information.
+# --------------------------------------------------------------------------------------------
+
+from abc import ABC, abstractmethod
+
+
+class AbstractBaseEventsHandler(ABC):
+    def __init__(self):
+        super().__init__()
+
+    @abstractmethod
+    def parse_message(self, message):
+        pass
