+++ azext_iot/operations/central.py
@@ -120,24 +120,24 @@ def _events3_runner(
         cmd, timeout, properties, enqueued_time, repair, yes
     )
 
-    from azext_iot.operations.events3 import _builders, _events
+    from azext_iot.monitor.builders import central_target_builder
+    from azext_iot.monitor.handlers import CommonHandler
+    from azext_iot.monitor.utility import generate_on_start_string
+    from azext_iot.monitor import telemetry
 
-    event_hub_targets = _builders.EventTargetBuilder().build_central_event_hub_target(
+    on_start_string = generate_on_start_string(device_id=device_id, pnp_context=None)
+
+    targets = central_target_builder.build_central_event_hub_targets(
         cmd, app_id, central_api_uri
     )
-    executorTargets = []
+    [target.add_consumer_group(consumer_group) for target in targets]
 
-    for target in event_hub_targets:
-        executorTargets.append(_events.executorData(target, consumer_group))
+    handler = CommonHandler(device_id=device_id, properties=properties, output=output)
 
-    _events.nExecutor(
-        executorTargets,
-        enqueued_time=enqueued_time,
-        properties=properties,
+    telemetry.start_multiple_monitors(
+        targets=targets,
+        enqueued_time_utc=enqueued_time,
+        on_start_string=on_start_string,
+        on_message_received=handler.parse_message,
         timeout=timeout,
-        device_id=device_id,
-        output=output,
-        validate_messages=validate_messages,
-        simulate_errors=simulate_errors,
-        central_device_provider=central_device_provider,
     )
