+++ azext_iot/operations/hub.py
@@ -1892,9 +1892,9 @@ def iot_c2d_message_send(
         if user_msg_expiry < now_in_milli:
             raise CLIError("Message expiry time utc is in the past!")
 
-    from azext_iot.operations.events3 import _events
+    from azext_iot.monitor import event
 
-    msg_id, errors = _events.send_c2d_message(
+    msg_id, errors = event.send_c2d_message(
         target=target,
         device_id=device_id,
         data=data,
@@ -2232,22 +2232,34 @@ def _iot_hub_monitor_events(
         cmd, hub_name, resource_group_name, include_events=True, login=login
     )
 
-    from azext_iot.operations.events3 import _builders, _events
+    from azext_iot.monitor.builders import hub_target_builder
+    from azext_iot.monitor.handlers import CommonHandler
+    from azext_iot.monitor.telemetry import start_single_monitor
+    from azext_iot.monitor.utility import generate_on_start_string
 
-    eventHubTarget = _builders.EventTargetBuilder().build_iot_hub_target(target)
+    target = hub_target_builder.EventTargetBuilder().build_iot_hub_target(target)
+    target.add_consumer_group(consumer_group)
 
-    _events.executor(
-        eventHubTarget,
-        consumer_group=consumer_group,
-        enqueued_time=enqueued_time,
-        properties=properties,
-        timeout=timeout,
+    on_start_string = generate_on_start_string(
+        device_id=device_id, pnp_context=pnp_context
+    )
+
+    handler = CommonHandler(
         device_id=device_id,
-        output=output,
-        content_type=content_type,
         devices=device_ids,
-        interface_name=interface_name,
         pnp_context=pnp_context,
+        interface_name=interface_name,
+        content_type=content_type,
+        properties=properties,
+        output=output,
+    )
+
+    start_single_monitor(
+        target=target,
+        enqueued_time_utc=enqueued_time,
+        on_start_string=on_start_string,
+        on_message_received=handler.parse_message,
+        timeout=timeout,
     )
 
 
@@ -2278,9 +2290,9 @@ def iot_hub_distributed_tracing_update(
 
 
 def _iot_hub_monitor_feedback(target, device_id, wait_on_id):
-    from azext_iot.operations.events3 import _events
+    from azext_iot.monitor import event
 
-    _events.monitor_feedback(
+    event.monitor_feedback(
         target=target, device_id=device_id, wait_on_id=wait_on_id, token_duration=3600
     )
